more refactoring and fixes
