+++ Makefile
@@ -122,8 +122,11 @@ endif
 	nimbus_portal_client \
 	fluffy \
 	nimbus_verified_proxy \
+	nimbus_verified_proxy_test \
 	libverifproxy \
-	libverifproxy-test \
+	libverifproxy_test \
+	nimbus_verified_proxy_wasm \
+	nimbus_verified_proxy_wasm_debug \
 	external_sync \
 	test \
 	test-reproducibility \
@@ -183,8 +186,6 @@ all: | $(TOOLS) nimbus nimbus_execution_client
 
 # must be included after the default target
 -include $(BUILD_SYSTEM_DIR)/makefiles/targets.mk
-
-include nimbus_verified_proxy/bindings/wasm/build_wasm.mk
 
 # "-d:release" cannot be added to config.nims
 
@@ -355,29 +356,52 @@ portal_bridge: | build deps
 # Builds the nimbus_verified_proxy
 nimbus_verified_proxy: | build deps
 	echo -e $(BUILD_MSG) "build/$@" && \
-		$(ENV_SCRIPT) nim nimbus_verified_proxy $(NIM_PARAMS) nimbus.nims
+		$(ENV_SCRIPT) nim c -o:build/$@ $(NIM_PARAMS) nimbus_verified_proxy/nimbus_verified_proxy.nim
 
 # builds and runs the nimbus_verified_proxy test suite
-nimbus-verified-proxy-test: | build deps
-	$(ENV_SCRIPT) nim nimbus_verified_proxy_test $(NIM_PARAMS) nimbus.nims
+nimbus_verified_proxy_test: | build deps
+	echo -e $(BUILD_MSG) "build/$@" && \
+		$(ENV_SCRIPT) nim c -r $(NIM_PARAMS) nimbus_verified_proxy/tests/all_proxy_tests.nim
 
 # Shared library for verified proxy
 
 libverifproxy: | build deps
-	+ echo -e $(BUILD_MSG) "build/$@" && \
-		$(ENV_SCRIPT) nim --version && \
-		echo $(NIM_PARAMS) && \
-		$(ENV_SCRIPT) nim c --app:staticlib --mm:orc -d:"libp2p_pki_schemes=secp256k1" --noMain:on -d:disable_libbacktrace --out:$(VERIF_PROXY_OUT_PATH)/$@.$(STATICLIBEXT) $(NIM_PARAMS) nimbus_verified_proxy/bindings/c/setup.nim
-	cp nimbus_verified_proxy/bindings/c/verifproxy.h $(VERIF_PROXY_OUT_PATH)/
-	echo -e $(BUILD_END_MSG) "build/$@"
+	echo -e $(BUILD_MSG) "build/$@" && \
+		$(ENV_SCRIPT) nim c \
+		--out:$(VERIF_PROXY_OUT_PATH)/$@.$(STATICLIBEXT) \
+		$(NIM_PARAMS) \
+		nimbus_verified_proxy/library/verifproxy.nim
+	cp nimbus_verified_proxy/library/verifproxy.h $(VERIF_PROXY_OUT_PATH)/
 
-libverifproxy-test: $(VERIF_PROXY_OUT_PATH)/libverifproxy.$(STATICLIBEXT)
+libverifproxy_test: libverifproxy
 	$(CC) -I$(VERIF_PROXY_OUT_PATH) -L$(VERIF_PROXY_OUT_PATH) \
 		-Wno-incompatible-pointer-types \
-		-o build/libverifproxy-test \
-		tests/bindings/c/test_api.c \
+		-o build/$@ \
+		tests/library/test_api.c \
 		-lverifproxy -lstdc++ $(VERIFPROXY_LDFLAGS)
-	./build/libverifproxy-test
+	./build/$@
+
+nimbus_verified_proxy_wasm: | build deps
+	@mkdir -p $(CURDIR)/build/$@
+	echo -e $(BUILD_MSG) "build/$@" && \
+		$(ENV_SCRIPT) nim c \
+		-d:emscripten \
+		-d:release \
+		-d:disable_libbacktrace \
+		-o:"$(CURDIR)/build/$@/verifproxy_wasm.js" \
+		nimbus_verified_proxy/library/bindings/wasm/verifproxy_wasm.nim
+	cp nimbus_verified_proxy/library/bindings/wasm/wasm_glue.js $(CURDIR)/build/$@/
+
+nimbus_verified_proxy_wasm_debug: | build deps
+	@mkdir -p $(CURDIR)/build/$@
+	echo -e $(BUILD_MSG) "build/$@" && \
+		$(ENV_SCRIPT) nim c \
+		-d:emscripten \
+		-d:debug \
+		-d:disable_libbacktrace \
+		-o:"$(CURDIR)/build/$@/verifproxy_wasm.js" \
+		nimbus_verified_proxy/library/bindings/wasm/verifproxy_wasm.nim
+	cp nimbus_verified_proxy/library/bindings/wasm/wasm_glue.js $(CURDIR)/build/$@/
 
 # Stateless related targets
