--- nimbus_verified_proxy/bindings/c/nim.cfg
@@ -1,19 +1,0 @@
--u:metrics
--d:setupForeignThreadGc=yes
-
-@if emscripten:
-  --cc:clang
-  --clang.exe:emcc
-  --clang.linkerexe:emcc
-  --cpu:wasm32
-  --os:linux
-  --mm:arc
-  --noMain:on
-  -d:noSignalHandler
-  -d:useMalloc
-  -d:disableMarchNative
-  -d:disable_libbacktrace
-  "-d:chronicles_timestamps=UnixTime"
-  --debugger:off
-  --passC:"-Wno-incompatible-pointer-types"
-@end
