+++ nimbus_verified_proxy/library/bindings/wasm/nim.cfg
@@ -1,0 +1,43 @@
+# nimbus_verified_proxy
+# Copyright (c) 2026 Status Research & Development GmbH
+# Licensed and distributed under either of
+#   * MIT license (license terms in the root directory or at https://opensource.org/licenses/MIT).
+#   * Apache v2 license (license terms in the root directory or at https://www.apache.org/licenses/LICENSE-2.0).
+# at your option. This file may not be copied, modified, or distributed except according to those terms.
+
+-u:metrics
+-d:setupForeignThreadGc=yes
+
+@if emscripten:
+  --cc:clang
+  --clang.exe:emcc
+  --clang.linkerexe:emcc
+  --cpu:wasm32
+  --os:linux
+  --mm:arc
+  --noMain:on
+  -d:noSignalHandler
+  -d:useMalloc
+  -d:disableMarchNative
+  -d:chronicles_timestamps=UnixTime
+  --debugger:off
+  --passC:"-Wno-incompatible-pointer-types"
+  --path:"shims"
+  --passL:"-sEXPORTED_FUNCTIONS=[\"_wasmStart\",\"_wasmStop\",\"_wasmCall\",\"_wasmDeliverExecutionTransport\",\"_wasmDeliverBeaconTransport\",\"_wasmFreeString\",\"_wasmProcessTasks\",\"_malloc\",\"_free\",\"_wasmExecCtxUrl\",\"_wasmExecCtxName\",\"_wasmExecCtxParams\",\"_wasmBeaconCtxUrl\",\"_wasmBeaconCtxEndpoint\",\"_wasmBeaconCtxParams\"]"
+  --passL:"-sEXPORTED_RUNTIME_METHODS=[\"ccall\",\"cwrap\",\"UTF8ToString\",\"stringToNewUTF8\",\"addFunction\",\"removeFunction\"]"
+  --passL:"-sALLOW_MEMORY_GROWTH=1"
+  --passL:"-sALLOW_TABLE_GROWTH=1"
+  --passL:"-sINITIAL_MEMORY=512MB"
+  --passL:"-sSTACK_SIZE=8MB"
+  --passL:"-sABORTING_MALLOC=1"
+  --passL:"-sMODULARIZE=1"
+  --passL:"-sEXPORT_NAME=VerifProxyModule"
+  --passL:"-sEXPORT_ES6=1"
+  @if release:
+    --passL:"-O1 -flto -Wl,--error-limit=0"
+  @end
+  @if debug:
+    --passL:"-O0 -g4 --profiling-funcs -Wl,--error-limit=0"
+    --passL:"-sASSERTIONS=2 -sSTACK_OVERFLOW_CHECK=2"
+  @end
+@end
