+++ nimbus_verified_proxy/library/bindings/wasm/verifproxy_wasm.nim
@@ -1,0 +1,13 @@
+# nimbus_verified_proxy
+# Copyright (c) 2026 Status Research & Development GmbH
+# Licensed and distributed under either of
+#   * MIT license (license terms in the root directory or at https://opensource.org/licenses/MIT).
+#   * Apache v2 license (license terms in the root directory or at https://www.apache.org/licenses/LICENSE-2.0).
+# at your option. This file may not be copied, modified, or distributed except according to those terms.
+
+when not defined(emscripten):
+  {.error: "verifproxy_wasm.nim must be compiled with -d:emscripten".}
+
+{.compile: "wasm_wrapper.c".}
+
+include ../../verifproxy
