+++ nimbus_verified_proxy/library/bindings/wasm/wasm_wrapper.c
@@ -10,7 +10,7 @@
 #include <stdint.h>
 #include <stdio.h>
 #include <emscripten.h>
-#include "../c/verifproxy.h"
+#include "../../verifproxy.h"
 
 static Context *g_ctx = NULL;
