+++ nimbus_verified_proxy/library/nim.cfg
@@ -1,0 +1,5 @@
+-u:metrics
+-d:setupForeignThreadGc=yes
+--app:staticlib
+--noMain:on
+-d:libp2p_pki_schemes=secp256k1
